upgrade b4.5

+++ MeadowInvaders/MeadowApp.cs
@@ -94,7 +94,7 @@ namespace MeadowInvaders
                 chipSelectPin: null,
                 dcPin: Device.Pins.D01,
                 resetPin: Device.Pins.D00,
-                width: (uint)displayWidth, height: (uint)displayHeight);
+                width: displayWidth, height: displayHeight);
 
             graphics = new GraphicsLibrary(display);
